#50 Add main application structure with navigation and page setup for Dashboard, Todo, and Diary
