+++ frontend_flutter/lib/pages/dashboard.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Dashboard extends StatelessWidget {
+  const Dashboard({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Center(child: Text("Dashboard!!"))
+        ],
+      ),
+    );
+  }
+}
