+++ frontend_flutter/lib/pages/todo.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Todo extends StatelessWidget {
+  const Todo({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Center(child: Text("TODO Page!"))
+        ],
+      ),
+    );
+  }
+}
